Improve iOS folder structure
