--- ios/KaMPKitiOS/BreedDetails/BreedDetailsNavController.swift
@@ -1,9 +1,0 @@
-//
-//  BreedDetailsNavController.swift
-//  KaMPKitiOS
-//
-//  Created by Bartłomiej Pedryc on 16/06/2023.
-//  Copyright © 2023 Touchlab. All rights reserved.
-//
-
-import Foundation
