+++ ios/KaMPKitiOS/BreedDetails/BreedDetailsViewModel.swift
@@ -1,16 +1,15 @@
 //
-//  BreedDetailsScreen.swift
+//  BreedDetailsViewModel.swift
 //  KaMPKitiOS
 //
-//  Created by Bartłomiej Pedryc on 07/06/2023.
+//  Created by Bartłomiej Pedryc on 05/07/2023.
 //  Copyright © 2023 Touchlab. All rights reserved.
 //
 
 import Combine
-import SwiftUI
+import Foundation
 import shared
 import KMPNativeCoroutinesCombine
-import Foundation
 
 class BreedDetailsViewModel: ObservableObject {
     private var viewModelDelegate: BreedDetailsViewModelDelegate
@@ -43,39 +42,5 @@ class BreedDetailsViewModel: ObservableObject {
 
     deinit {
         viewModelDelegate.clear()
-    }
-}
-
-struct BreedDetailsScreen: View {
-    @StateObject
-    var viewModel: BreedDetailsViewModel
-
-    var body: some View {
-        BreedDetailsContent(
-            breedName: viewModel.detailsState.breed.name,
-            isBreedFavorite: viewModel.detailsState.breed.favorite,
-            onFavoriteClick: { viewModel.onFavoriteClick() }
-        )
-        .onAppear(perform: {
-            viewModel.activate()
-        })
-        .onDisappear(perform: {
-            viewModel.deactivate()
-        })
-    }
-}
-
-struct BreedDetailsContent: View {
-    var breedName: String
-    var isBreedFavorite: Bool
-    var onFavoriteClick: () -> Void
-    var body: some View {
-        HStack {
-            Text(breedName)
-            Button(action: onFavoriteClick) {
-                Image(systemName: (!isBreedFavorite) ? "heart" : "heart.fill")
-                    .padding(4.0)
-            }
-        }
     }
 }
