--- ios/KaMPKitiOS/Breeds/BreedsNavController.swift
@@ -1,13 +1,0 @@
-//
-//  BreedsNavController.swift
-//  KaMPKitiOS
-//
-//  Created by Bartłomiej Pedryc on 16/06/2023.
-//  Copyright © 2023 Touchlab. All rights reserved.
-//
-
-import Foundation
-
-func buildBreedsNavController(navController: BreedsNavController) -> UIViewController {
-    let viewModel = Kotlin
-}
